+++ lib/truemetrics_flutter_sdk.dart
@@ -3,12 +3,16 @@ import 'package:truemetrics_flutter_sdk/truemetrics_config.dart';
 import 'truemetrics_flutter_sdk_platform_interface.dart';
 import 'upload_statistics.dart';
 import 'sensor_statistics.dart';
+import 'configuration.dart';
+import 'sensor_info.dart';
 
 export 'truemetrics_config.dart';
 export 'truemetrics_state.dart';
 export 'truemetrics_error.dart';
 export 'upload_statistics.dart';
 export 'sensor_statistics.dart';
+export 'configuration.dart';
+export 'sensor_info.dart';
 
 class TruemetricsFlutterSdk {
   Future<bool?> isInitialized() {
@@ -19,11 +23,13 @@ class TruemetricsFlutterSdk {
     StateChangeCallback? onStateChange,
     ErrorCallback? onError,
     PermissionsCallback? onPermissionsRequired,
+    ConfigChangeCallback? onConfigChange,
   }) {
     TruemetricsFlutterSdkPlatform.instance.setStatusListener(
       onStateChange: onStateChange,
       onError: onError,
       onPermissionsRequired: onPermissionsRequired,
+      onConfigChange: onConfigChange,
     );
   }
 
@@ -55,6 +61,34 @@ class TruemetricsFlutterSdk {
     return TruemetricsFlutterSdkPlatform.instance.getDeviceId();
   }
 
+  // --- Recording State APIs ---
+
+  Future<bool> isRecordingInProgress() {
+    return TruemetricsFlutterSdkPlatform.instance.isRecordingInProgress();
+  }
+
+  Future<bool> isRecordingStopped() {
+    return TruemetricsFlutterSdkPlatform.instance.isRecordingStopped();
+  }
+
+  Future<int> getRecordingStartTime() {
+    return TruemetricsFlutterSdkPlatform.instance.getRecordingStartTime();
+  }
+
+  // --- Sensor APIs ---
+
+  Future<void> setAllSensorsEnabled(bool enabled) {
+    return TruemetricsFlutterSdkPlatform.instance.setAllSensorsEnabled(enabled);
+  }
+
+  Future<bool> getAllSensorsEnabled() {
+    return TruemetricsFlutterSdkPlatform.instance.getAllSensorsEnabled();
+  }
+
+  Future<List<SensorInfo>> getSensorInfo() {
+    return TruemetricsFlutterSdkPlatform.instance.getSensorInfo();
+  }
+
   // --- Statistics APIs ---
 
   Future<UploadStatistics?> getUploadStatistics() {
@@ -63,6 +97,12 @@ class TruemetricsFlutterSdk {
 
   Future<List<SensorStatistics>> getSensorStatistics() {
     return TruemetricsFlutterSdkPlatform.instance.getSensorStatistics();
+  }
+
+  // --- Configuration APIs ---
+
+  Future<TruemetricsConfiguration?> getActiveConfig() {
+    return TruemetricsFlutterSdkPlatform.instance.getActiveConfig();
   }
 
   // --- Metadata Template APIs ---